+++ fatiando/inversion/hyper_param.py
@@ -72,7 +72,7 @@ class LCurve(OptimizerMixin):
     >>> from fatiando.mesher import SquareMesh
     >>> from fatiando.seismic import ttime2d, srtomo
     >>> from fatiando.inversion import Smoothness2D, LCurve
-    >>> from fatiando import utils
+    >>> from fatiando import utils, gridder
     >>> area = (0, 2, 0, 2)
     >>> shape = (10, 10)
     >>> model = SquareMesh(area, shape)
@@ -90,9 +90,13 @@ class LCurve(OptimizerMixin):
            [  4.,   4.,   4.,   4.,   4.,   4.,   4.,   4.,   4.,   4.],
            [  4.,   4.,   4.,   4.,   4.,   4.,   4.,   4.,   4.,   4.]])
     >>> model.addprop('vp', vp.ravel())
-    >>> src_loc = utils.random_points(area, 30, seed=0)
-    >>> rec_loc = utils.circular_points(area, 20, random=True, seed=0)
-    >>> srcs, recs = utils.connect_points(src_loc, rec_loc)
+    >>> src_loc_x, src_loc_y = gridder.scatter(area, 30, seed=0)
+    >>> src_loc = numpy.transpose([src_loc_x, src_loc_y])
+    >>> rec_loc_x, rec_loc_y = gridder.circular_scatter(area, 20,
+    ...                                                 random=True, seed=0)
+    >>> rec_loc = numpy.transpose([rec_loc_x, rec_loc_y])
+    >>> srcs = [src for src in src_loc for _ in rec_loc]
+    >>> recs = [rec for _ in src_loc for rec in rec_loc]
     >>> tts = ttime2d.straight(model, 'vp', srcs, recs)
     >>> tts = utils.contaminate(tts, 0.01, percent=True, seed=0)