+++ gallery/5-seismic/srtomo_regularized.py
@@ -18,7 +18,7 @@ import matplotlib.pyplot as plt
 from fatiando.mesher import SquareMesh
 from fatiando.seismic import ttime2d, srtomo
 from fatiando.inversion import Smoothness2D, Damping, TotalVariation2D
-from fatiando import utils
+from fatiando import utils, gridder
 
 # First, we'll create a simple model with a high velocity square in the middle
 area = (0, 500000, 0, 500000)
@@ -31,9 +31,13 @@ model.addprop('vp', vel.ravel())
 # Make some noisy travel time data using straight-rays
 # Set the random seed so that points are the same every time we run this script
 seed = 0
-src_loc = utils.random_points(area, 80, seed=seed)
-rec_loc = utils.circular_points(area, 30, random=True, seed=seed)
-srcs, recs = utils.connect_points(src_loc, rec_loc)
+src_loc_x, src_loc_y = gridder.scatter(area, 80, seed=seed)
+src_loc = np.transpose([src_loc_x, src_loc_y])
+rec_loc_x, rec_loc_y = gridder.circular_scatter(area, 30,
+                                                random=True, seed=seed)
+rec_loc = np.transpose([rec_loc_x, rec_loc_y])
+srcs = [src for src in src_loc for _ in rec_loc]
+recs = [rec for _ in src_loc for rec in rec_loc]
 tts = ttime2d.straight(model, 'vp', srcs, recs)
 # Use 2% random noise to corrupt the data
 tts = utils.contaminate(tts, 0.02, percent=True, seed=seed)